+++ src/core/widget_store.rs
@@ -16,51 +16,35 @@
 use crate::core::point::*;
 use crate::widget::widget::*;
 
-use opengl_graphics::Texture;
-use gl::types::GLuint;
-
+use opengl_graphics::GlGraphics;
 use piston_window::*;
 
-/// This is a container object, used for storing the `Widget` trait object, and the parent-child
-/// relationship for the added `Widget`.  Only the `widget` is public.
+/// This is a container object, used for storing the `Widget` trait object, and the parent
+/// relationship for the added `Widget`.  Only the `widget` is public.  `Widget` objects do not
+/// need to have a child relationship, only parent objects are traversed.  A parent object of 0, or
+/// itself, indicates that the parent is self.
 pub struct WidgetContainer {
     /// The `Widget` trait object being stored.
     pub widget: Box<dyn Widget>,
 
-    /// This `Widget`'s assigned ID.
+    /// This `Widget`'s assigned ID.  These IDs are auto-assigned.
     widget_id: i32,
 
     /// The parent ID.
     parent_id: i32,
 }
 
-/// This structure contains a window and its corresponding onscreen widgets.  These objects
-/// are stored in the Pushrod main loop.
-pub struct PushrodWindow {
-    /// A `piston_window::PistonWindow` object.
-    pub window: PistonWindow,
-
-    /// A vector list of Boxed `PushrodWidget` trait objects.
+/// This is the `WidgetStore`, which is used to store `Widget` objects for a `Pushrod`
+/// management object.
+pub struct WidgetStore {
+    /// A vector list of Boxed `WidgetContainer` objects.
     pub widgets: Vec<WidgetContainer>,
-
-    /// Texture buffer for the window, used by OpenGL, stored in heap.
-    texture_buf: Box<Vec<u8>>,
-
-    /// The texture against which objects may be drawn.
-    pub texture: Texture,
-
-    /// Framebuffer Object ID, stored for drawing on the texture.
-    pub fbo: GLuint,
 }
 
-/// Implementation for a new `PushrodWindow`.  When a new `PushrodWindow` is added to the
-/// managed window stack in the Pushrod main loop, this object is created to store its
-/// components.
-impl PushrodWindow {
-    /// Constructor, takes a managed `PistonWindow` from the `piston_window` crate.  Adds a top-level
-    /// widget to the list that is a white container widget.  This is the base for all other widgets
-    /// tht will be added to the window.
-    pub fn new(window: PistonWindow) -> Self {
+/// Implementation of the `WidgetStore`.
+impl WidgetStore {
+    /// Creates a new `WidgetStore`.
+    pub fn new() -> Self {
         let mut widgets_list: Vec<WidgetContainer> = Vec::new();
         let mut base_widget = BaseWidget::new();
 
@@ -72,11 +56,7 @@ impl PushrodWindow {
         });
 
         Self {
-            window,
             widgets: widgets_list,
-            texture_buf: Box::new(vec![0u8; 1]),
-            texture: Texture::empty(&TextureSettings::new()).unwrap(),
-            fbo: 0,
         }
     }
 
@@ -86,31 +66,6 @@ impl PushrodWindow {
     /// `pushrod::core::main` loop, but it can be handled programmatically if required.
     pub fn handle_resize(&mut self, width: u32, height: u32) {
         eprintln!("[Resize] W={} H={}", width, height);
-        self.texture_buf = Box::new(vec![0u8; width as usize * height as usize]);
-        self.texture = Texture::from_memory_alpha(&self.texture_buf, width, height,
-        &TextureSettings::new()).unwrap();
-
-        // I hate this code.  However, this does prepare a texture so that an image can be
-        // drawn on it.  Since it's in memory, it means that the texture only gets recreated once
-        // per resize.
-        unsafe {
-            let mut fbos: [GLuint; 1] = [0];
-
-            gl::GenFramebuffers(1, fbos.as_mut_ptr());
-            self.fbo = fbos[0];
-
-            gl::BindFramebuffer(gl::FRAMEBUFFER, self.fbo);
-            gl::FramebufferTexture2D(gl::FRAMEBUFFER,
-                gl::COLOR_ATTACHMENT0, gl::TEXTURE_2D,
-                self.texture.get_id(),
-                0);
-        }
-    }
-
-    /// Prepares the initial buffers for drawing.  Do not call more than once.
-    pub fn prepare_buffers(&mut self) {
-        let draw_size = self.window.draw_size();
-        self.handle_resize(draw_size.width as u32, draw_size.height as u32);
     }
 
     /// Invalidates all widgets in the window.  This is used to force a complete refresh of the
@@ -118,6 +73,16 @@ impl PushrodWindow {
     /// care, as this is an expensive operation.
     pub fn invalidate_all_widgets(&mut self) {
         self.widgets.iter_mut().for_each(|x| x.widget.invalidate());
+    }
+
+    /// Indicates whether or not any `Widget`s in the `WidgetStore` have been invalidated and need
+    /// to be repainted.
+    pub fn needs_repaint(&mut self) -> bool {
+        self.widgets
+            .iter_mut()
+            .map(|x| x.widget.is_invalidated())
+            .find(|x| x == &true)
+            .unwrap_or(false)
     }
 
     /// Adds a UI `Widget` to this window.  `Widget` objects that are added using this method will
@@ -196,6 +161,72 @@ impl PushrodWindow {
         }
 
         found_id
+    }
+
+    /// Recursive draw object: paints objects in order of appearance on the screen.  This does not
+    /// account for object depth, but it is implied that objects' parents are displayed in stacking
+    /// order.  Therefore, the parent is drawn first, then sibling, and other siblings.  This draw
+    /// function is used by the `Pushrod` main loop, and is meant to be called in a `draw_2d`
+    /// closure.
+    pub fn draw(&mut self, widget_id: i32, c: Context, g: &mut G2d) {
+        let parents_of_widget = self.get_children_of(widget_id);
+
+        if parents_of_widget.len() == 0 {
+            return;
+        }
+
+        for pos in 0..parents_of_widget.len() {
+            let paint_id = parents_of_widget[pos];
+            let paint_widget = &mut self.widgets[paint_id as usize];
+
+            if &paint_widget.widget.is_invalidated() == &true {
+                // Implementation of auto-clipping.  Clips the object's drawing area.
+                if paint_widget.widget.get_autoclip() {
+                    let trans = c.transform.trans(
+                        paint_widget.widget.get_origin().x as f64,
+                        paint_widget.widget.get_origin().y as f64,
+                    );
+                    let viewport = c.viewport.unwrap();
+                    let scale_x = viewport.draw_size[0] as f64 / viewport.window_size[0];
+                    let scale_y = viewport.draw_size[1] as f64 / viewport.window_size[1];
+
+                    let clip_rect = [
+                        ((paint_widget.widget.get_origin().x as f64 + viewport.rect[0] as f64)
+                            * scale_x) as u32,
+                        ((paint_widget.widget.get_origin().y as f64 + viewport.rect[1] as f64)
+                            * scale_y) as u32,
+                        (paint_widget.widget.get_size().w as f64 * scale_x) as u32,
+                        (paint_widget.widget.get_size().h as f64 * scale_y) as u32,
+                    ];
+
+                    let vp = Viewport {
+                        rect: [
+                            paint_widget.widget.get_origin().x as i32 + viewport.rect[0],
+                            paint_widget.widget.get_origin().y as i32 + viewport.rect[1],
+                            paint_widget.widget.get_size().w as i32,
+                            paint_widget.widget.get_size().h as i32,
+                        ],
+                        draw_size: viewport.draw_size,
+                        window_size: viewport.window_size,
+                    };
+
+                    let clipped = Context {
+                        viewport: Some(vp),
+                        view: c.view,
+                        transform: trans,
+                        draw_state: c.draw_state.scissor(clip_rect),
+                    };
+
+                    &paint_widget.widget.draw(clipped, g);
+                } else {
+                    &paint_widget.widget.draw(c, g);
+                }
+            }
+
+            if parents_of_widget[pos] != widget_id {
+                self.draw(paint_id, c, g);
+            }
+        }
     }
 
     /// Callback to `mouse_entered` for a `Widget` by ID.