+++ src/widget/timer_widget.rs
@@ -57,7 +57,7 @@ impl TimerWidget {
             enabled: true,
             initiated: time_ms(),
             timeout: 0,
-            timeout_function: Box::new(|| { }),
+            timeout_function: Box::new(|| {}),
         }
     }
 
@@ -104,18 +104,9 @@ impl TimerWidget {
 /// ```no_run
 /// # use piston_window::*;
 /// # use pushrod::core::point::*;
-/// # use pushrod::core::window::*;
 /// # use pushrod::widget::widget::*;
 /// # use pushrod::widget::timer_widget::*;
 /// # fn main() {
-/// #   let opengl = OpenGL::V3_2;
-/// #   let mut pushrod_window: PushrodWindow = PushrodWindow::new(
-/// #       WindowSettings::new("Pushrod Window", [640, 480])
-/// #           .opengl(opengl)
-/// #           .build()
-/// #           .unwrap_or_else(|error| panic!("Failed to build PistonWindow: {}", error)),
-/// #   );
-/// #
 ///    let mut timer_widget = TimerWidget::new();
 ///
 ///    timer_widget.set_timeout(60000);
@@ -155,7 +146,7 @@ impl Widget for TimerWidget {
 
     /// Does not draw anything - only calls the timer `tick()` function to increment the
     /// timer.
-    fn draw(&mut self, _context: Context, _graphics: &mut GlGraphics) {
+    fn draw(&mut self, _context: Context, _graphics: &mut G2d) {
         self.tick();
     }
 }